Wrap simulation in a loop
Doing so will allow the program to re-run many times as long as the
specified condition is met, which is if playAgain stays true.

Right now, we can ask the user if they want to continue playing, but
the program does not yet have a way of interpreting that response as
a trigger on whether the loop should continue or end.